Extract main method into new PetGame class
