+++ src/trunky/pet/Main.java
@@ -1,0 +1,7 @@
+package trunky.pet;
+
+public class Main {
+    public static void main(String[] args) {
+        PetGame.getGame().run();
+    }
+}
